+++ Runtime/Visuals/CurrencyBinding.cs
@@ -1,0 +1,29 @@
+﻿using System.ComponentModel;
+using UnityEngine;
+
+
+namespace THEBADDEST.VirtualCurrencySystem
+{
+
+
+	public abstract class CurrencyBinding : MonoBehaviour
+	{
+
+		[SerializeField] protected CurrencyType currencyTypeName = CurrencyType.Coin;
+
+		protected virtual void OnDestroy()
+		{
+			VirtualCurrencyService.OnValueChangeUnregister(currencyTypeName, ChangeEffect);
+		}
+
+		protected abstract void ChangeEffect(object sender, PropertyChangedEventArgs args);
+
+		protected virtual void Start()
+		{
+			VirtualCurrencyService.OnValueChangeRegister(currencyTypeName, ChangeEffect);
+		}
+
+	}
+
+
+}